pyproject.toml
[project]
name = "top-servers-discord-bot"
version = "0.1.0"
description = "TopServersDiscord Discord bot (stats collector)"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "discord.py==2.7.1",
    "pydantic==2.13.1",
    "pydantic-settings==2.7.0",
    "clickhouse-connect==0.8.10",
    "valkey==6.0.2",
    "httpx==0.28.1",
    "uvloop==0.21.0; sys_platform != 'win32'",
    "structlog==25.1.0",
    "prometheus-client==0.21.1",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.4",
    "pytest-asyncio==0.25.0",
    "dpytest==0.7.0",
    "ruff==0.8.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["bot"]

[tool.ruff]
line-length = 100
target-version = "py314"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
